Adapt Pyright and Intelephense servers to solidlsp
