+++ src/solidlsp/language_servers/intelephense/intelephense.py
@@ -2,27 +2,25 @@
 Provides PHP specific instantiation of the LanguageServer class using Intelephense.
 """
 
-import asyncio
 import json
-import shutil
 import logging
 import os
-import subprocess
 import pathlib
-from contextlib import asynccontextmanager
+import shutil
+import subprocess
 from time import sleep
-from typing import AsyncIterator
 
 from overrides import override
 
-from multilspy.multilspy_logger import MultilspyLogger
-from multilspy.language_server import LanguageServer
-from multilspy.lsp_protocol_handler.server import ProcessLaunchInfo
 from multilspy.lsp_protocol_handler.lsp_types import DefinitionParams, InitializeParams
+from multilspy.lsp_protocol_handler.server import ProcessLaunchInfo
 from multilspy.multilspy_config import MultilspyConfig
+from multilspy.multilspy_logger import MultilspyLogger
 from multilspy.multilspy_utils import PlatformUtils, PlatformId
+from solidlsp.ls import SolidLanguageServer
 
-class Intelephense(LanguageServer):
+
+class Intelephense(SolidLanguageServer):
     """
     Provides PHP specific instantiation of the LanguageServer class using Intelephense.
     """
@@ -109,7 +107,6 @@ class Intelephense(LanguageServer):
             ProcessLaunchInfo(cmd=intelephense_cmd, cwd=repository_root_path),
             "php"
         )
-        self.server_ready = asyncio.Event()
         self.request_id = 0
 
     def _get_initialize_params(self, repository_absolute_path: str) -> InitializeParams:
@@ -136,16 +133,15 @@ class Intelephense(LanguageServer):
 
         return d
 
-    @asynccontextmanager
-    async def start_server(self) -> AsyncIterator["Intelephense"]:
+    def _start_server(self):
         """Start Intelephense server process"""
-        async def register_capability_handler(params):
+        def register_capability_handler(params):
             return
 
-        async def window_log_message(msg):
+        def window_log_message(msg):
             self.logger.log(f"LSP: window/logMessage: {msg}", logging.INFO)
 
-        async def do_nothing(params):
+        def do_nothing(params):
             return
 
         self.server.on_request("client/registerCapability", register_capability_handler)
@@ -153,48 +149,44 @@ class Intelephense(LanguageServer):
         self.server.on_notification("$/progress", do_nothing)
         self.server.on_notification("textDocument/publishDiagnostics", do_nothing)
 
-        async with super().start_server():
-            self.logger.log("Starting Intelephense server process", logging.INFO)
-            await self.server.start()
-            initialize_params = self._get_initialize_params(self.repository_root_path)
+        self.logger.log("Starting Intelephense server process", logging.INFO)
+        self.server.start()
+        initialize_params = self._get_initialize_params(self.repository_root_path)
 
-            self.logger.log(
-                "Sending initialize request from LSP client to LSP server and awaiting response",
-                logging.INFO,
-            )
-            init_response = await self.server.send.initialize(initialize_params)
-            self.logger.log(
-                "After sent initialize params",
-                logging.INFO,
-            )
-            
-            # Verify server capabilities
-            assert "textDocumentSync" in init_response["capabilities"]
-            assert "completionProvider" in init_response["capabilities"]
-            assert "definitionProvider" in init_response["capabilities"]
+        self.logger.log(
+            "Sending initialize request from LSP client to LSP server and awaiting response",
+            logging.INFO,
+        )
+        init_response = self.server.send.initialize(initialize_params)
+        self.logger.log(
+            "After sent initialize params",
+            logging.INFO,
+        )
 
-            self.server.notify.initialized({})
-            self.completions_available.set()
+        # Verify server capabilities
+        assert "textDocumentSync" in init_response["capabilities"]
+        assert "completionProvider" in init_response["capabilities"]
+        assert "definitionProvider" in init_response["capabilities"]
 
-            # Intelephense server is typically ready immediately after initialization
-            self.server_ready.set()
-            await self.server_ready.wait()
+        self.server.notify.initialized({})
+        self.completions_available.set()
 
-            yield self
-            
+        # Intelephense server is typically ready immediately after initialization
+        # TODO: This is probably incorrect; the server does send an initialized notification, which we could wait for!
+
     @override
     # For some reason, the LS may need longer to process this, so we just retry
-    async def _send_references_request(self, relative_file_path: str, line: int, column: int):
+    def _send_references_request(self, relative_file_path: str, line: int, column: int):
         # TODO: The LS doesn't return references contained in other files if it doesn't sleep. This is
         #   despite the LS having processed requests already. I don't know what causes this, but sleeping
         #   one second helps. It may be that sleeping only once is enough but that's hard to reliably test.
         # May be related to the time it takes to read the files or something like that.
         # The sleeping doesn't seem to be needed on all systems
         sleep(1)
-        return await super()._send_references_request(relative_file_path, line, column)
+        return super()._send_references_request(relative_file_path, line, column)
     
     @override
-    async def _send_definition_request(self, definition_params: DefinitionParams):
+    def _send_definition_request(self, definition_params: DefinitionParams):
         # TODO: same as above, also only a problem if the definition is in another file
         sleep(1)
-        return await super()._send_definition_request(definition_params)
+        return super()._send_definition_request(definition_params)
